refactored Authorization Componet to be multistep form
